Move data utils tests into the more applicable file for where these functions are actually defined.
